Fixed some warnings, fixed console bug

+++ Libraries/libpd/x_libpd_mod_utils.h
@@ -26,7 +26,7 @@ t_pd* libpd_creategraphonparent(t_canvas* cnv, int x, int y);
 
 void libpd_finishremove(t_canvas* cnv);
 void libpd_removeobj(t_canvas* cnv, t_gobj* obj);
-void libpd_renameobj(t_canvas* cnv, t_gobj* obj, const char* buf, int bufsize);
+void libpd_renameobj(t_canvas* cnv, t_gobj* obj, const char* buf, size_t bufsize);
 void libpd_moveobj(t_canvas* cnv, t_gobj* obj, int x, int y);
